move getTooltipWithoutTrendline from trendline to tooltips
