+++ src/utils/trendline.js
@@ -102,25 +102,4 @@ function trendlineSlope(trendlineValues) {
   return (lastValue - firstValue) / trendlineValues.length;
 }
 
-/**
- * Returns the string value for a tooltip that excludes information about
- * trendline data points.
- * Appends the `units` string to the data yLabel value if provided.
- */
-function getTooltipWithoutTrendline(tooltipItem, data, units) {
-  const { label } = data.datasets[tooltipItem.datasetIndex];
-  if (label === "trendline") return "";
-  let tooltipLabel = tooltipItem.yLabel.toString();
-  if (units) {
-    tooltipLabel = tooltipLabel.concat(units);
-  }
-
-  return tooltipLabel;
-}
-
-export {
-  generateTrendlineDataset,
-  getTooltipWithoutTrendline,
-  trendlineSlope,
-  trendlineData,
-};
+export { generateTrendlineDataset, trendlineSlope, trendlineData };
